+++ ansible_collections/arista/cvp/plugins/module_utils/device_tools.py
@@ -1434,16 +1434,14 @@ class CvDeviceTools(object):
         results = []
         for device in user_inventory.devices:
             result_data = CvApiResult(
-                action_name="{}_to_{}".format(device.fqdn, *device.container)
+                action_name="{0}_to_{1}".format(device.fqdn, device.container)
             )
             if device.system_mac is not None:
                 new_container_info = self.get_container_info(
                     container_name=device.container
                 )
                 if new_container_info is None:
-                    error_message = "The target container '{0}' for the device '{1}' does not exist on CVP.".format(
-                        device.container, device.fqdn
-                    )
+                    error_message = f"The target container '{device.container}' for the device '{device.fqdn}' does not exist on CVP."
                     MODULE_LOGGER.error(error_message)
                     self.__ansible.fail_json(msg=error_message)
 
@@ -1471,21 +1469,17 @@ class CvDeviceTools(object):
                                 create_task=True,
                             )
                         except CvpApiError:
-                            error_message = (
-                                "Error to move device {} to container {}".format(
-                                    device.fqdn, *device.container
-                                )
-                            )
+                            error_message = f"Error to move device {device.fqdn} to container {device.container}"
                             MODULE_LOGGER.error(error_message)
                             self.__ansible.fail_json(msg=error_message)
                         else:
-                            if resp["data"]["status"] == "success":
+                            if resp and resp['data']['status'] == 'success':
                                 result_data.changed = True
                                 result_data.success = True
                                 result_data.taskIds = resp["data"][Api.task.TASK_IDS]
 
                     result_data.add_entry(
-                        "{}-{}".format(device.fqdn, *device.container)
+                        "{0}-{1}".format(device.fqdn, device.container)
                     )
             results.append(result_data)
         return results
@@ -1534,6 +1528,9 @@ class CvDeviceTools(object):
                 or current_container_info[Api.generic.NAME]
                 == Api.container.UNDEFINED_CONTAINER_ID
             ):
+                MODULE_LOGGER.debug(
+                    "The device is in undefined container"
+                )
                 continue
 
             # GET IMAGE BUNDLE
@@ -1554,6 +1551,7 @@ class CvDeviceTools(object):
 
             if device.image_bundle is not None:
                 if (
+                    current_image_bundle and
                     device.image_bundle
                     == current_image_bundle[Api.generic.IMAGE_BUNDLE_NAME]
                     and current_image_bundle[Api.image.TYPE] == "netelement"
@@ -1590,11 +1588,7 @@ class CvDeviceTools(object):
                             "Error image bundle %s not found", str(device.image_bundle)
                         )
                         self.__ansible.fail_json(
-                            msg="Error applying bundle to device"
-                            + device.fqdn
-                            + ": "
-                            + str(device.image_bundle)
-                            + "not found"
+                            msg=f"Error applying bundle to device {device.fqdn}: {str(device.image_bundle)} not found"
                         )
 
                     MODULE_LOGGER.debug(
@@ -1626,13 +1620,10 @@ class CvDeviceTools(object):
                                 "Error applying bundle to device: %s", str(catch_error)
                             )
                             self.__ansible.fail_json(
-                                msg="Error applying bundle to device"
-                                + device.fqdn
-                                + ": "
-                                + catch_error
+                                msg=f"Error applying bundle to device {device.fqdn}: {str(catch_error)}"
                             )
                         else:
-                            if resp["data"]["status"] == "success":
+                            if resp and resp["data"]["status"] == "success":
                                 result_data.changed = True
                                 result_data.success = True
                                 result_data.taskIds = resp["data"][Api.task.TASK_IDS]
@@ -1732,13 +1723,10 @@ class CvDeviceTools(object):
                                 str(catch_error),
                             )
                             self.__ansible.fail_json(
-                                msg="Error removing bundle from device"
-                                + device.fqdn
-                                + ": "
-                                + catch_error
+                                msg=f"Error removing bundle from device {device.fqdn}: {str(catch_error)}"
                             )
                         else:
-                            if resp["data"]["status"] == "success":
+                            if resp and resp["data"]["status"] == "success":
                                 result_data.changed = True
                                 result_data.success = True
                                 result_data.taskIds = resp["data"][Api.task.TASK_IDS]
@@ -2250,7 +2238,7 @@ class CvDeviceTools(object):
                     MODULE_LOGGER.error("Error resetting device")
                     self.__ansible.fail_json(msg="Error resetting device")
                 else:
-                    if resp["data"]["status"] == "success":
+                    if resp and resp["data"]["status"] == "success":
                         result_data.changed = True
                         result_data.success = True
                         result_data.taskIds = resp["data"][Api.task.TASK_IDS]